+++ include/metl/metl.hpp
@@ -16,6 +16,7 @@
 #include "metl/crc32.hpp"
 #include "metl/crc8.hpp"
 #include "metl/delegate.hpp"
+#include "metl/detail/construct.hpp"
 #include "metl/endian.hpp"
 #include "metl/event_dispatcher.hpp"
 #include "metl/expected.hpp"